+++ stc_game/Assets/Scripts/Models/DatabaseModels/BaseUser.cs
@@ -7,7 +7,6 @@ public class BaseUser
     public string Name;
     public string UserAuthenticationId;
     public List<CharacterBaseModel> Characters;
-    public GameState GameState;
     
     public BaseUser()
     {
@@ -28,5 +27,34 @@ public class BaseUser
         Name = name;
         UserAuthenticationId = userAuthenticationId;
         Characters = new List<CharacterBaseModel>();
+    }
+
+    public void SelectCharacter(HeroClass heroClass)
+    {
+        GetActiveCharacter().IsActive = false;
+        GetCharacter(heroClass).IsActive = true;
+    }
+
+    private CharacterBaseModel GetCharacter(HeroClass heroClass)
+    {
+
+        return Characters.Find(character => character.HeroClass == heroClass);
+    }
+
+    public CharacterBaseModel GetActiveCharacter()
+    {
+        if (Characters.Count == 0)
+        {
+            InstantiateCharacters();
+        }
+        return Characters.Find(character => character.IsActive == true);
+    }
+
+    private void InstantiateCharacters()
+    {
+        CharacterBaseModel warrior = new CharacterBaseModel("Warrior", HeroClass.Warrior, true);
+        CharacterBaseModel mage = new CharacterBaseModel("Mage", HeroClass.Mage);
+        Characters.Add(warrior);
+        Characters.Add(mage);
     }
 }
