+++ stc_game/Assets/Scripts/Models/DatabaseModels/GameState.cs
@@ -1,10 +1,42 @@
 ﻿using System;
 using System.Collections.Generic;
+using Devdog.QuestSystemPro;
 
 [Serializable]
 public class GameState
 {
+    public class QuestState
+    {
+        public QuestsContainer QuestContainer = new QuestsContainer();
+    }
+
+    public class InventoryState
+    {
+        public List<string> Items;
+    }
+
+    public class EquipmentState
+    {
+        public List<string> Equipment;
+    }
+
+    public class StatsState
+    {
+        public Stats Stats;
+    }
+
     public string UserId;
-    public string CharacterId;
+    public HeroClass HeroClass;
+
     // Add GameState save fields here
+
+    // Quest Manager / Quest Data
+    // Inventory 
+    // Equipment
+    // Stats
+
 }
+
+
+
+
