+++ stc_game/Assets/Scripts/Models/DatabaseModels/User.cs
@@ -32,33 +32,4 @@ public class User: BaseUser
             return false;
         }
     }
-
-    public void SelectCharacter(HeroClass heroClass)
-    {
-        GetActiveCharacter().IsActive = false;
-        GetCharacter(heroClass).IsActive = true;
-    }
-
-    private CharacterBaseModel GetCharacter(HeroClass heroClass)
-    {
-        
-        return Characters.Find(character => character.HeroClass == heroClass);
-    }
-
-    public CharacterBaseModel GetActiveCharacter()
-    {
-        if (Characters.Count == 0)
-        {
-            InstantiateCharacters();
-        }
-        return Characters.Find(character => character.IsActive == true);
-    }
-
-    private void InstantiateCharacters()
-    {
-        CharacterBaseModel warrior = new CharacterBaseModel("Warrior", HeroClass.Warrior, true);
-        CharacterBaseModel mage = new CharacterBaseModel("Mage", HeroClass.Mage);
-        Characters.Add(warrior);
-        Characters.Add(mage);
-    }
 }
